Move Newton actuator manager authoring test to isaaclab_tasks
The TestManagerBasedSceneNewtonActuatorAuthoring test imports isaaclab_tasks
(G1FlatEnvCfg) so it belongs in source/isaaclab_tasks/test/, not in
source/isaaclab_newton/test/ which should not depend on isaaclab_tasks.
